silence compile warnings
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@1954 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ test/testcases/varn_intf.f
@@ -249,7 +249,7 @@
           call check(err, 'In nfmpi_get_varn_int_all: ')
 
  997      format(A,I2,A,I2,A,I2)
-          do i=1, w_len
+          do i=1, INT(w_len)
              if (buffer(i) .NE. rank) then
                  print 997, "Error: expecting buffer(",i,")=",rank,
      +                      " but got", buffer(i)
